requirements.txt
flask
werkzeug
pypdf2
langchain
langchain-community
langchain-google-genai
google-generativeai
sentence-transformers
faiss-cpu
python-dotenv
gunicorn